refactor(web/mcp): extract AuthBandFrame from AuthActionBand
